+++ spotdl/providers/audio/piped.py
@@ -10,10 +10,9 @@ from typing import Any, Dict, List, Optional
 import requests
 from yt_dlp import YoutubeDL
 
-from spotdl.download.config import DownloadConfig
 from spotdl.providers.audio.base import ISRC_REGEX, AudioProvider, YTDLLogger
 from spotdl.types.result import Result
-from spotdl.utils.config import get_temp_path
+from spotdl.utils.config import GlobalConfig, get_temp_path
 from spotdl.utils.formatter import args_to_ytdlp_options
 
 __all__ = ["Piped"]
@@ -149,7 +148,7 @@ class Piped(AudioProvider):
         piped_data = requests.get(
             f"https://pipedapi.kavin.rocks/streams/{url_id}",
             timeout=10,
-            proxies=DownloadConfig.get_parameter("proxies"),
+            proxies=GlobalConfig.get_parameter("proxies"),
         ).json()
 
         yt_dlp_json = {
