+++ spotdl/providers/audio/sliderkz.py
@@ -7,9 +7,9 @@ from typing import Any, Dict, List
 
 import requests
 
-from spotdl.download.config import DownloadConfig
 from spotdl.providers.audio.base import AudioProvider
 from spotdl.types.result import Result
+from spotdl.utils.config import GlobalConfig
 
 __all__ = ["SliderKZ"]
 
@@ -50,7 +50,7 @@ class SliderKZ(AudioProvider):
                     url="https://slider.kz/vk_auth.php?q=" + search_term,
                     headers=HEADERS,
                     timeout=5,
-                    proxies=DownloadConfig.get_parameter("proxies"),
+                    proxies=GlobalConfig.get_parameter("proxies"),
                 )
 
                 # Check if the response is valid
