+++ spotdl/providers/lyrics/azlyrics.py
@@ -34,7 +34,7 @@ class AzLyrics(LyricsProvider):
 
         self.x_code = js_code[start_index : start_index + end_index]
 
-    def get_results(self, name: str, artists: List[str], **kwargs) -> Dict[str, str]:
+    def get_results(self, name: str, artists: List[str], **_) -> Dict[str, str]:
         """
         Returns the results for the given song.
