+++ spotdl/utils/metadata.py
@@ -40,8 +40,8 @@ from mutagen.id3._specs import Encoding
 from mutagen.mp4 import MP4Cover
 from mutagen.wave import WAVE
 
-from spotdl.download.config import DownloadConfig
 from spotdl.types.song import Song
+from spotdl.utils.config import GlobalConfig
 from spotdl.utils.formatter import to_ms
 
 logger = logging.getLogger(__name__)
@@ -275,7 +275,7 @@ def embed_cover(audio_file, song: Song, encoding: str):
         cover_data = requests.get(
             song.cover_url,
             timeout=10,
-            proxies=DownloadConfig.get_parameter("proxies"),
+            proxies=GlobalConfig.get_parameter("proxies"),
         ).content
     except Exception:
         return audio_file
